Merge original repository

+++ tests/TestCase.php
@@ -2,9 +2,9 @@
 
 namespace Hedii\LaravelDateRange\Tests;
 
-use File;
-use Illuminate\Foundation\Application;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Foundation\Application;
+use Illuminate\Support\Facades\File;
 use Orchestra\Testbench\TestCase as Orchestra;
 
 class TestCase extends Orchestra
@@ -14,14 +14,14 @@ class TestCase extends Orchestra
      */
     protected $testModel;
 
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
 
         $this->setUpDatabase($this->app);
     }
 
-    protected function getEnvironmentSetUp($app)
+    protected function getEnvironmentSetUp($app): void
     {
         $this->initializeDirectory($this->getTempDirectory());
 
@@ -33,29 +33,38 @@ class TestCase extends Orchestra
         ]);
     }
 
-    protected function setUpDatabase(Application $app)
+    protected function setUpDatabase(Application $app): void
     {
-        file_put_contents($this->getTempDirectory() . '/database.sqlite', null);
+        file_put_contents($this->getTempDirectory() . '/' . $this->getTempSqliteFile(), null);
 
         $app['db']->connection()->getSchemaBuilder()->create('test_models', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name')->nullable();
             $table->string('other_field')->nullable();
             $table->string('url')->nullable();
+            $table->timestamp('logged_in_at')->nullable();
             $table->timestamps();
         });
     }
 
-    protected function initializeDirectory(string $directory)
+    protected function initializeDirectory(string $directory): void
     {
-        if (File::isDirectory($directory)) {
-            File::deleteDirectory($directory);
+        if (! File::exists($directory . '/' . $this->getTempSqliteFile())) {
+            if (File::isDirectory($directory)) {
+                File::deleteDirectory($directory);
+            }
+            File::makeDirectory($directory);
         }
-        File::makeDirectory($directory);
+
     }
 
-    public function getTempDirectory()
+    public function getTempDirectory(): string
     {
         return __DIR__ . '/temp';
+    }
+
+    public function getTempSqliteFile(): string
+    {
+        return 'database.sqlite';
     }
 }
